+++ tools/monitoring/s2_gap_filler.py
@@ -11,12 +11,14 @@ import click
 from odc.aws import s3_fetch, s3_client
 from odc.aws.queue import get_queue, publish_messages
 
-from tools.monitoring.utils import (
+from tools.utils.utils import (
     find_latest_report,
     read_report,
     split_list_equally,
+    send_slack_notification,
+    setup_logging,
+    slack_url,
 )
-from tools.monitoring.utils import send_slack_notification, setup_logging, slack_url
 
 PRODUCT_NAME = "s2_l2a"
 SOURCE_REGION = "us-west-2"
